requirements.txt
# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3

# Multi-tenancy
django-tenants==3.5.0
django-tenant-schemas==3.0.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# Database
psycopg2-binary==2.9.7
redis==5.0.1

# Async tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Payment processing
stripe==7.7.0
razorpay==1.3.0

# AI/ML (optional)
openai==1.3.5
transformers==4.35.2

# Monitoring & Logging
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0

# Development
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3
